requirements.txt
SpeechRecognition>=0.10.0
pyttsx3>=2.90
pyaudio>=0.2.14
openai>=1.0.0
pyautogui>=0.3.0
Pillow>=10.0.0
pytesseract>=0.3.10
psutil>=5.9.0
schedule>=1.2.0
requests>=2.31.0
python-dotenv>=1.0.0
plyer>=2.1.0
colorama>=0.4.6
pyperclip>=1.8.3
chardet>=5.1.0
PyPDF2>=3.0.0
